Added the bee model and texture. Cleaned and commented some code.
This looks pretty good, only cam goes clipping on occasion. Need to add some eye candy like walls/rail in form of forest or canyon on sides. Add in obstacles and resource pickups. A distance fog shader would help a lot too. Texture the track and perhaps smoothly alternate the height of the side of the track to make it concave/convex but then we need normals too.